requirements.txt
Django==4.2.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
rasterio==1.3.8
opencv-python==4.8.0.76
scikit-image==0.21.0
scipy==1.11.1
Pillow==10.0.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
celery==5.3.1
redis==4.6.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0